requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
gevent==23.9.1
gevent-websocket==0.10.1
pymavlink==2.4.39
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1 